+++ libs/asset_type_apigw.py
@@ -60,8 +60,9 @@ class APIGW(AssetType):
             self.authorization.add_auth_type(auth_type)
         self.location.region = region_name
         self.backend_endpoint = []
+        self.domain_name = None
 
-    def report(self, report, brief=False):
+    def report(self, report, brief=False, with_fpkey=False):
         """
         Add an asset with only relevent informations
         """
@@ -73,10 +74,12 @@ class APIGW(AssetType):
                 'AuthorizationTypes': self.authorization.types,
                 'Backend Endpoints': self.backend_endpoint
             }
+            if self.domain_name:
+                asset_report['Domain Name'] = self.domain_name
             if self.public:
                 asset_report['PubliclyAccessible'] = '[red]True[/red]'
             if self.security_issues:
-                self.update_audit_report(asset_report)
+                self.update_audit_report(asset_report, with_fpkey)
         if 'APIGW' not in report[self.location.region]:
             report[self.location.region]['APIGW'] = { self.name: asset_report }
             return report
@@ -126,7 +129,7 @@ def get_raw_data(raw_data, authorizations, boto_session, cache, _):
             raw_data['ag_rest_api_raw'][rest_api['id']] = []
             resources = ag_client.get_resources(restApiId=rest_api['id'])['items']
             for resource in resources:
-                if 'resourceMethods' not in resource:
+                if 'resourceMethods' not in resource or 'GET' not in resource['resourceMethods']:
                     continue
                 raw_data['ag_rest_api_raw'][rest_api['id']].append(
                     ag_client.get_integration(
@@ -159,6 +162,17 @@ def get_raw_data(raw_data, authorizations, boto_session, cache, _):
             'list_functions')['Functions']
     except botocore.exceptions.ClientError as err_msg:
         authorizations['apigw'] = False
+
+    # Get domain API mapping
+    all_domains = cache.get(
+        'ag_get_domain_names',
+        ag_client,
+        'get_domain_names')['items']
+    raw_data['ag_domain_raw'] = {}
+    for domain in all_domains:
+        raw_data['ag_domain_raw'][domain['domainName']] = \
+            ag_client.get_base_path_mappings(domainName=domain['domainName'])
+
     return raw_data, authorizations
 
 @log_me('Scanning API Gateway...')
@@ -167,51 +181,66 @@ def parse_raw_data(assets, authorizations, raw_data, public_only, boto_session, 
     Parsing the raw data to extracts assets,
     enrich the assets list and add a 'False' in authorizations in case of errors
     """
-    for apigw in raw_data['ag_raw']:
-        asset = cache.get_asset(f'APIGW_{apigw["name"]}')
-        if asset is None:
-            is_public = 'REGIONAL' in apigw['endpointConfiguration']['types']
-            if public_only and not is_public:
-                continue
-            asset = APIGW(
-                apigw['name'],
-                apigw['id'],
-                boto_session.region_name,
-                [apigw['apiKeySource']],
-                public=is_public)
-            for rest_api in raw_data['ag_rest_api_raw'][apigw['id']]:
-                asset.backend_endpoint.append(
-                    get_lambda_name(rest_api['uri']))
-            cache.save_asset(f'APIGW_{apigw["name"]}', asset)
-        if asset is not None and name_filter.lower() in asset.name.lower():
-            assets.append(asset)
-    for apigw in raw_data['agv2_raw']:
-        asset = cache.get_asset(f'APIGW_{apigw["Name"]}')
-        if asset is None:
-            authorization_types = []
-            try:
-                for route in raw_data['agv2_client'].get_routes(ApiId=apigw['ApiId'])['Items']:
-                    authorization_types.append(route['AuthorizationType'])
-            except botocore.exceptions.ClientError:
-                authorizations['apigw'] = False
-            asset = APIGW(
-                apigw['Name'],
-                apigw['ApiId'],
-                boto_session.region_name,
-                authorization_types,
-                public=True)
-            for rest_api in raw_data['agv2_rest_api_raw'][apigw['ApiId']]:
-                asset.backend_endpoint.append(
-                    get_lambda_name(rest_api['IntegrationUri']))
-            cache.save_asset(f'APIGW_{apigw["Name"]}', asset)
-        if search_filter_in(asset, name_filter):
-            assets.append(asset)
-    for lambda_fun in raw_data['lambda_raw']:
-        asset = cache.get_asset(f'LAMBDA_{lambda_fun["FunctionName"]}')
-        if asset is None:
-            authorization_types = []
-            asset = Lambda(lambda_fun['FunctionName'])
-            cache.save_asset(f'LAMBDA_{lambda_fun["FunctionName"]}', asset)
-        if search_filter_in(asset, name_filter):
-            assets.append(asset)
+    if 'ag_raw' in raw_data:
+        for apigw in raw_data['ag_raw']:
+            asset = cache.get_asset(f'APIGW_{apigw["name"]}')
+            if asset is None:
+                is_public = 'REGIONAL' in apigw['endpointConfiguration']['types']
+                if public_only and not is_public:
+                    continue
+                asset = APIGW(
+                    apigw['name'],
+                    apigw['id'],
+                    boto_session.region_name,
+                    [apigw['apiKeySource']],
+                    public=is_public)
+                for rest_api in raw_data['ag_rest_api_raw'][apigw['id']]:
+                    if 'uri' not in rest_api:
+                        rest_api['uri'] = 'unknown'
+                    asset.backend_endpoint.append(
+                        get_lambda_name(rest_api['uri']))
+                for domain in raw_data['ag_domain_raw']:
+                    if 'items' in raw_data['ag_domain_raw'][domain] and \
+                        raw_data['ag_domain_raw'][domain]['items'][0] and \
+                        raw_data['ag_domain_raw'][domain]['items'][0]['restApiId'] == apigw['id']:
+                        asset.domain_name = domain
+                cache.save_asset(f'APIGW_{apigw["name"]}', asset)
+            if asset is not None and name_filter.lower() in asset.name.lower():
+                assets.append(asset)
+    if 'agv2_raw' in raw_data:
+        for apigw in raw_data['agv2_raw']:
+            asset = cache.get_asset(f'APIGW_{apigw["Name"]}')
+            if asset is None:
+                authorization_types = []
+                try:
+                    for route in raw_data['agv2_client'].get_routes(ApiId=apigw['ApiId'])['Items']:
+                        authorization_types.append(route['AuthorizationType'])
+                except botocore.exceptions.ClientError:
+                    authorizations['apigw'] = False
+                asset = APIGW(
+                    apigw['Name'],
+                    apigw['ApiId'],
+                    boto_session.region_name,
+                    authorization_types,
+                    public=True)
+                for rest_api in raw_data['agv2_rest_api_raw'][apigw['ApiId']]:
+                    asset.backend_endpoint.append(
+                        get_lambda_name(rest_api['IntegrationUri']))
+                for domain in raw_data['ag_domain_raw']:
+                    if 'items' in raw_data['ag_domain_raw'][domain] and \
+                        raw_data['ag_domain_raw'][domain]['items'][0] and \
+                        raw_data['ag_domain_raw'][domain]['items'][0]['restApiId'] == apigw['ApiId']:
+                        asset.domain_name = domain
+                cache.save_asset(f'APIGW_{apigw["Name"]}', asset)
+            if search_filter_in(asset, name_filter):
+                assets.append(asset)
+    if 'lambda_raw' in raw_data:
+        for lambda_fun in raw_data['lambda_raw']:
+            asset = cache.get_asset(f'LAMBDA_{lambda_fun["FunctionName"]}')
+            if asset is None:
+                authorization_types = []
+                asset = Lambda(lambda_fun['FunctionName'])
+                cache.save_asset(f'LAMBDA_{lambda_fun["FunctionName"]}', asset)
+            if search_filter_in(asset, name_filter):
+                assets.append(asset)
     return assets, authorizations