+++ libs/asset_type_elb.py
@@ -27,7 +27,7 @@ class ELB(AssetType):
         self.dns_record = None
         self.targets = []
 
-    def report(self, report, brief=False):
+    def report(self, report, brief=False, with_fpkey=False):
         """
         Add an asset with only relevent informations
         """
@@ -44,7 +44,7 @@ class ELB(AssetType):
             if self.dns_record:
                 asset_report['DnsRecord'] = self.dns_record
             if self.security_issues:
-                self.update_audit_report(asset_report)
+                self.update_audit_report(asset_report, with_fpkey)
         if 'ELB' not in report[self.location.region][self.location.vpc][self.location.subnet]:
             report[self.location.region][self.location.vpc][self.location.subnet]['ELB'] = \
                 { self.name: asset_report }