+++ monitoring/aws_lambda_patrowl/aws_tower_launcher.py
@@ -10,20 +10,62 @@ Written by Nicolas BEGUIER (nicolas.beguier@adevinta.com)
 from configparser import ConfigParser
 import json
 import logging
+import os
 import sys
 import boto3
 
 # Third party library imports
 sys.path.append('package')
+from patrowl4py.api import PatrowlManagerApi
+
+# Own library and config files
+from libs.patrowl import add_in_assetgroup
 
 # pylint: disable=logging-fstring-interpolation
 
 LOGGER = logging.getLogger('aws-tower-launcher')
 
-VERSION = '4.6.0'
+VERSION = '4.3.0'
+
+PATROWL = {}
+PATROWL['api_token'] = os.environ['PATROWL_APITOKEN']
+PATROWL['assetgroup_dev'] = int(os.environ['PATROWL_DEV_ASSETGROUP'])
+PATROWL['assetgroup_pre'] = int(os.environ['PATROWL_PRE_ASSETGROUP'])
+PATROWL['assetgroup_pro'] = int(os.environ['PATROWL_PRO_ASSETGROUP'])
+PATROWL['private_endpoint'] = os.environ['PATROWL_PRIVATE_ENDPOINT']
+PATROWL['public_endpoint'] = os.environ['PATROWL_PUBLIC_ENDPOINT']
 
 LOGGER = logging.getLogger('aws-tower')
 
+PATROWL_API = PatrowlManagerApi(
+    url=PATROWL['private_endpoint'],
+    auth_token=PATROWL['api_token']
+)
+
+def organize_assetgroups(config):
+    """
+    Organize all assetgroups by adding automatically all assets
+    This will be done once, not in every call_lambda.
+    A Patrowl call is long and add latencies in every lambdas...
+    """
+    patrowl_assets = PATROWL_API.get_assets()
+    assetgroup = {}
+    assetgroup['dev'] = []
+    assetgroup['pre'] = []
+    assetgroup['pro'] = []
+    for asset in patrowl_assets:
+        for profile in config.sections():
+            if not is_config_ok(config, profile):
+                continue
+            aws_account_name = profile.split()[1]
+            if asset['name'].startswith(f'[{aws_account_name}]'):
+                assetgroup[config[profile]['env']].append(asset['id'])
+    for env in assetgroup:
+        add_in_assetgroup(
+            PATROWL_API,
+            PATROWL[f'assetgroup_{env}'],
+            assetgroup[env])
+        LOGGER.warning(f'Add these IDs in {env}: {assetgroup[env]}')
 
 def is_config_ok(config, profile):
     """
@@ -62,6 +104,7 @@ def main():
     """
     config = ConfigParser(strict=False)
     config.read('config/lambda.config')
+    organize_assetgroups(config)
     # A lambda per profile
     for profile in config.sections():
         if not is_config_ok(config, profile):